build 64 bit snoop
struct timeval --> struct timeval32

+++ usr/src/cmd/cmd-inet/usr.sbin/snoop/snoop_ntp.c
@@ -210,7 +210,7 @@ interpret_ntp(int flags, struct ntpdata *ntp_pkt, int fraglen)
 				 */
 				break;
 		    }
-		    (void) sprintf(get_line((char *)ntp->keyid -
+		    (void) sprintf(get_line((char *)(uintptr_t)ntp->keyid -
 			dlc_header, 1),
 			"Key ID  = %8lu", ntohl(ntp->keyid));
